+++ azure/packages/azure-client/src/utils.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
-import type { ConfigTypes, IConfigProviderBase } from "@fluidframework/core-interfaces";
 import type {
 	AzureConnectionConfig,
 	AzureLocalConnectionConfig,
@@ -26,19 +25,3 @@ export function isAzureLocalConnectionConfig(
 ): connectionConfig is AzureLocalConnectionConfig {
 	return connectionConfig.type === "local";
 }
-
-/**
- * Wrapper {@link IConfigProviderBase} which allows for specifying feature gates if not present in the wrapped
- * provider.
- *
- * @param defaults - default feature gate configs to be used if not specified by the original provider
- * @param original - the original config provider
- * @returns The value for the requested feature gate from the original provider and if not present,
- * from the specified feature gates
- */
-export const wrappedConfigProviderWithDefaults = (
-	defaults: Record<string, ConfigTypes>,
-	original: IConfigProviderBase | undefined,
-): IConfigProviderBase => ({
-	getRawConfig: (name: string): ConfigTypes => original?.getRawConfig(name) ?? defaults[name],
-});
