+++ packages/utils/telemetry-utils/src/index.ts
@@ -9,6 +9,7 @@ export {
 	mixinMonitoringContext,
 	IConfigProvider,
 	loggerToMonitoringContext,
+	wrapConfigProviderWithDefaults,
 } from "./config";
 export {
 	DataCorruptionError,
